expose-loader(options): New option to handle ESM (Harmony) modules
Before this change the expose loader could only handle exposing the one thing exported from module.exports. This is great for modules that are written using RJS, but Harmony modules are an issue, since in the RJS representation of a Harmony module you might have named exports or the default export that you want to expose (without wanting to have to use .default every time). This adds an object syntax to the loader options which allows you to expose a single key of the RJS representation of the Harmony module instead of the whole thing.

Fixes #56